# docs/openapi.yaml
...
openapi: 3.0.2
info:
  title: Breeding Insight API
  description: API documentation for Breeding Insight
  termsOfService: http://swagger.io/terms/
  contact:
    email: np398@cornell.edu
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:8081/v1
...
paths:
...
  /programs/{programId}/traits/{traitId}:
    get:
      tags:
        - traits
      summary: Get a single trait by traitId
      description: Gets a single trait with full level of detail
      operationId: getTrait
      parameters:
        - name: programId
          in: path
          description: Id of program to get trait in
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - name: traitId
          in: path
          description: Id of trait to get trait for
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/singleTraitResponse'
              examples:
                singleTraitResponse:
                  $ref: '#/components/examples/singleTrait'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
              example: |-
                ERROR - 2018-10-08T20:15:11Z - Malformed JSON Request Object
                ERROR - 2018-10-08T20:15:11Z - Invalid query parameter
                ERROR - 2018-10-08T20:15:11Z - Required parameter is missing
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - Missing or expired authorization
                token
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - User does not have permission
                to perform this action
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - The requested object DbId is
                not found
...
  /programs/{programId}/trait-upload:
    put:
      tags:
        - uploads
      parameters:
        - name: programId
          in: path
          description: The id of the program
          required: true
          style: simple
          explode: false
          schema:
            type: string
      summary: Upload the trait csv/xls/xlsx file data
      description: Upload trait file data for the user making the request in the indicated program. Only one trait upload at a time can exist per user per program.
      operationId: putTraitUpload
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
            encoding:
              file:
                contentType: text/csv, application/vnd.ms-excel, application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/traitUploadResponse'
              examples:
                traitUploadResponse:
                  $ref: '#/components/examples/traitUploadData'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
              example: |-
                ERROR - 2018-10-08T20:15:11Z - Malformed JSON Request Object
                ERROR - 2018-10-08T20:15:11Z - Invalid query parameter
                ERROR - 2018-10-08T20:15:11Z - Required parameter is missing
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - Missing or expired authorization
                token
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - User does not have permission
                to perform this action
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - The requested object DbId is
                not found
        "422":
          description: Multi-error UnprocessableEntity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/validatorErrorResponse"
              example:
                rowErrors:
                  - rowIndex: 2
                    errors:
                      - column: "Trait level"
                        errorMessage: "Missing trait name"
                        httpStatus: "UNPROCESSABLE_ENTITY"
                        httpStatusCode: 422
                      - column: "Trait abbreviations"
                        errorMessage: "Abbreviations already exist"
                        httpStatus: "CONFLICT"
                        httpStatusCode: 409
                      - column: "Scale class"
                        errorMessage: "Missing scale class"
                        httpStatus: "UNPROCESSABLE_ENTITY"
                        httpStatusCode: 422
    get:
      tags:
        - uploads
      summary: Get the trait upload data
      description: Gets the trait upload data for the specified program and requesting user
      operationId: getTraitUpload
      parameters:
        - name: programId
          in: path
          description: Id of program to get trait upload in
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - in: query
          description: Field to sort the results by.
          name: sortField
          required: false
          schema:
            type: string
            enum: [name, abbreviations, mainAbbreviation, synonyms, level, status, methodDescription, methodClass, methodFormula, scaleName, scaleClass, scaleDecimalPlaces, scaleLowerLimit, scaleUpperLimit, scaleCategories, createdAt, updatedAt, createdByUserId, createdByUserName, updatedByUserId, updatedByUserName]
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
        - $ref: '#/components/parameters/sortOrder'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/traitUploadResponse'
              examples:
                traitUploadResponse:
                  $ref: '#/components/examples/traitUploadData'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
              example: |-
                ERROR - 2018-10-08T20:15:11Z - Malformed JSON Request Object
                ERROR - 2018-10-08T20:15:11Z - Invalid query parameter
                ERROR - 2018-10-08T20:15:11Z - Required parameter is missing
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - Missing or expired authorization
                token
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - User does not have permission
                to perform this action
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - The requested object DbId is
                not found
    delete:
      tags:
        - uploads
      parameters:
        - name: programId
          in: path
          description: Id of program to delete trait upload from
          required: true
          style: simple
          explode: false
          schema:
            type: string
      summary: Delete trait upload
      description: Delete trait upload for the specified program and requesting user
      operationId: deleteTraitUpload
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
              example: |-
                ERROR - 2018-10-08T20:15:11Z - Malformed JSON Request Object
                ERROR - 2018-10-08T20:15:11Z - Invalid query parameter
                ERROR - 2018-10-08T20:15:11Z - Required parameter is missing
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - Missing or expired authorization
                token
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - User does not have permission
                to perform this action
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
              example: ERROR - 2018-10-08T20:15:11Z - The requested object DbId is
                not found
  /programs/{programId}/trait-upload/search:
    post:
      tags:
        - uploads
      parameters:
        - name: programId
          in: path
          description: Id of program to get trait upload for
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - in: query
          description: Field to sort the results by.
          name: sortField
          required: false
          schema:
            type: string
            enum: [name, abbreviations, mainAbbreviation, synonyms, level, status, methodDescription, methodClass, methodFormula, scaleName, scaleClass, scaleDecimalPlaces, scaleLowerLimit, scaleUpperLimit, scaleCategories, createdAt, updatedAt, createdByUserId, createdByUserName, updatedByUserId, updatedByUserName]
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
        - $ref: '#/components/parameters/sortOrder'
      summary: Search trait upload
      description: Search trait upload using filters
      operationId: searchTraitUpload
      requestBody:
        description: Search request
        content:
          application/json:
            schema:
              type: object
              properties:
                filter:
                  type: array
                  items:
                    type: object
                    properties:
                      field:
                        type: string
                        enum: [name, abbreviations, mainAbbreviation, synonyms, level, status, methodDescription, methodClass, methodFormula, scaleName, scaleClass, scaleDecimalPlaces, scaleLowerLimit, scaleUpperLimit, scaleCategories, createdAt, updatedAt, createdByUserId, createdByUserName, updatedByUserId, updatedByUserName]
                      value:
                        type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/traitUploadResponse'
              examples:
                traitUploadResponse:
                  $ref: '#/components/examples/traitUploadData'
        '400':
          $ref: '#/components/responses/400BadRequest'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        "404":
          $ref: '#/components/responses/404NotFound'
  /programs/{programId}/trait-upload/{traitUploadId}:
    post:
      tags:
        - uploads
      parameters:
        - name: programId
          in: path
          description: Id of program to get trait upload for
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - name: traitUploadId
          in: path
          description: Id of the trait upload to save
          required: true
          style: simple
          explode: false
          schema:
            type: string
      summary: Confirm and save existing trait upload as traits
      description: Confirm and save existing trait upload as traits
      operationId: postTraitUpload
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/multipleTraitsResponse'
              examples:
                multipleTraitsResponse:
                  $ref: '#/components/examples/arrayOfTraits'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '404':
          $ref: '#/components/responses/404NotFound'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '422':
          description: Multi-error UnprocessableEntity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/validatorErrorResponse"
              example:
                rowErrors:
                  - rowIndex: 0
                    errors:
                      - column: "Trait level"
                        errorMessage: "Missing trait name"
                        httpStatus: "UNPROCESSABLE_ENTITY"
                        httpStatusCode: 422
                      - column: "Trait abbreviations"
                        errorMessage: "Abbreviations already exist"
                        httpStatus: "CONFLICT"
                        httpStatusCode: 409
                      - column: "Scale class"
                        errorMessage: "Missing scale class"
                        httpStatus: "UNPROCESSABLE_ENTITY"
                        httpStatusCode: 422
...
components:
  schemas:
    metadata:
      type: object
      properties:
        datafiles:
          type: array
          description: The datafiles key contains a list of file paths, which can
            be relative or complete URLs. These files contain additional information
            related to the returned object and can be retrieved by a subsequent call.
            The empty list should be returned if no additional data files are present.
          items:
            type: string

        pagination:
          $ref: '#/components/schemas/metadata_pagination'
        status:
          type: array
          description: The status field contains a list of informational status messages
            from the server. If no status is reported, an empty list should be returned.
            See Error Reporting for more information.
          items:
            $ref: '#/components/schemas/status'
...
    multipleTraitsResponse:
      required:
        - data
      type: object
      properties:
        data:
          type: array
          description: Array of traits
          items:
            $ref: '#/components/schemas/trait'
    singleTraitResponse:
      required:
        - metadata
        - result
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/metadata'
        result:
          $ref: '#/components/schemas/trait'
    trait:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: System id of the trait
        traitName:
          type: string
          description: Trait display name
        active:
          type: boolean
          description: Trait active status
        createdAt:
          type: string
          format: datetime
        updatedAt:
          type: string
          format: datetime
        createdByUser:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        updatedByUser:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
        traitClass:
          type: string
          example: "morphological"
        attribute:
          type: string
          description: Attribute the trait is describing
          example: height
        defaultValue:
          type: string
          description: Default value of the trait
        mainAbbreviation:
          type: string
        abbreviations:
          type: array
          items:
            type: string
        synonyms:
          type: array
          items:
            type: string
        programObservationLevel:
          type: object
          properties:
            id:
              type: string
              format: uuid
              description: System id of the observation level
            name:
              type: string
              description: Name of the observation level
        method:
          type: object
          properties:
            methodClass:
              type: string
            description:
              type: string
            formula:
              type: string
        scale:
          type: object
          properties:
            scaleName:
              type: string
            dataType:
              type: string
            decimalPlaces:
              type: integer
            categories:
              type: array
              items:
                type: object
                properties:
                  label:
                    type: string
                  value:
                    type: string
            validValueMax:
              type: integer
            validValueMin:
              type: integer
    traitUploadResponse:
      required:
        - metadata
        - result
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/metadata'
        result:
          $ref: '#/components/schemas/traitUpload'
...
    validatorErrorResponse:
      type: object
      properties:
        rowErrors:
          type: array
          items:
            type: object
            properties:
              rowIndex:
                type: integer
                description: "Row number in the uploaded file for the trait with errors."
              errors:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      description: "Field name that is the source of the error"
                    errorMessage:
                      type: string
                      description: "A description fo the error"
                    httpStatus:
                      type: string
                      description: "Text representation fo the error code"
                    httpStatusCode:
                      type: integer
                      description: "Http error code"
...
  responses:
    "400BadRequest":
      description: Bad Request
      content:
        application/json:
          schema:
            type: string
          example: |-
            ERROR - 2018-10-08T20:15:11Z - Malformed JSON Request Object
            ERROR - 2018-10-08T20:15:11Z - Invalid query parameter
            ERROR - 2018-10-08T20:15:11Z - Required parameter is missing
    "401Unauthorized":
      description: Unauthorized
      content:
        application/json:
          schema:
            type: string
          example: ERROR - 2018-10-08T20:15:11Z - Missing or expired authorization
            token
    "403Forbidden":
      description: Forbidden
      content:
        application/json:
          schema:
            type: string
          example: ERROR - 2018-10-08T20:15:11Z - User does not have permission to
            perform this action
    "404NotFound":
      description: Not Found
      content:
        application/json:
          schema:
            type: string
          example: ERROR - 2018-10-08T20:15:11Z - The requested object DbId is not
            found
  securitySchemes:
    orcid_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://sandbox.orcid.org/oauth/authorize
          scopes:
            write:bi: modify
            read:bi: read
    api_key:
      type: apiKey
      name: api_key
      in: header
  examples:
...
    arrayOfTraits:
      value:
        metadata:
          datafiles: []
          pagination:
            currentPage: 1
            pageSize: 1
            totalCount: 1
            totalPages: 1
          status: []
        result:
          data:
            [
            {
              id: "00385063-01c8-4afc-a684-3c84b90b2de9",
              traitName: "Test Trait",
              active: true,
              createdAt: "2020-06-04T11:05:48-04:00",
              updatedAt: "2020-06-04T11:05:48-04:00",
              traitClass: "morphological",
              attribute: "height",
              defaultValue: "defaultValue",
              entity: "stalk",
              mainAbbreviation: "t1",
              abbreviations: ["t1","t2"],
              synonyms: ["stalk height"],
              programObservationLevel:
                {
                  id: "71421c6a-19d4-4d82-858b-e3f70a9c7a02",
                  name: "Plant",
                  active: true,
                  createdAt: "2020-06-04T11:05:48-04:00",
                  updatedAt: "2020-06-04T11:05:48-04:00",
                },
              method:
                {
                  methodClass: "Counting"
                },
              scale:
                {
                  scaleName: "Test Scale",
                  dataType: "TEXT",
                  decimalPlaces: 3,
                  categories:
                    [
                    {
                      label: "test1",
                      value: "value1"
                    },
                    {
                      label: "test2",
                      value: "value2"
                    }
                    ]
                },
              programOntology:
                {
                  id: "a9712526-85ca-4a47-afe6-974ce4abf6a9",
                  programId: "c9caac57-3dcb-442b-b3e9-e01281fbf8bd",
                  createdAt: "2020-06-04T11:05:48-04:00",
                  updatedAt: "2020-06-04T11:05:48-04:00"
                },
              createdByUser:
                {
                  name: "Tester",
                  active: true
                },
              updatedByUser:
                {
                  name: "Tester",
                  active: true
                }
            }
            ]
    singleTrait:
      value:
        metadata:
          datafiles: []
          pagination:
            currentPage: 1
            pageSize: 1
            totalCount: 1
            totalPages: 1
          status: []
        result:
          {
            id: "00385063-01c8-4afc-a684-3c84b90b2de9",
            traitName: "Test Trait",
            active: true,
            createdAt: "2020-06-04T11:05:48-04:00",
            updatedAt: "2020-06-04T11:05:48-04:00",
            traitClass: "morphological",
            attribute: "height",
            defaultValue: "defaultValue",
            entity: "stalk",
            mainAbbreviation: "t1",
            abbreviations: ["t1","t2"],
            synonyms: ["stalk height"],
            programObservationLevel:
              {
                id: "71421c6a-19d4-4d82-858b-e3f70a9c7a02",
                name: "Plant",
                active: true,
                createdAt: "2020-06-04T11:05:48-04:00",
                updatedAt: "2020-06-04T11:05:48-04:00",
              },
            method:
              {
                methodClass: "Counting"
              },
            scale:
              {
                scaleName: "Test Scale",
                dataType: "TEXT",
                decimalPlaces: 3,
                categories:
                  [
                  {
                    label: "test1",
                    value: "value1"
                  },
                  {
                    label: "test2",
                    value: "value2"
                  }
                  ]
              },
            programOntology:
              {
                id: "a9712526-85ca-4a47-afe6-974ce4abf6a9",
                programId: "c9caac57-3dcb-442b-b3e9-e01281fbf8bd",
                createdAt: "2020-06-04T11:05:48-04:00",
                updatedAt: "2020-06-04T11:05:48-04:00"
              },
            createdByUser:
              {
                name: "Tester",
                active: true
              },
            updatedByUser:
              {
                name: "Tester",
                active: true
              }
          }
    traitUploadData:
      value:
        metadata:
          datafiles: []
          pagination:
            currentPage: 1
            pageSize: 1
            totalCount: 1
            totalPages: 1
          status: []
        result:
          type: "TRAIT"
          data:
            [
            {
              traitName: "Test Trait",
              active: true,
              programObservationLevel:
                {
                  name: "Plant",
                },
              method:
                {
                  methodClass: "Estimation",
                  description: "Observed severity of Powdery Mildew on leaves",
                  formula: "a^2 + b^2 = c^2"
                },
              scale:
                {
                  scaleName: "1-4 Parlier field response score",
                  dataType: "ORDINAL",
                  decimalPlaces: 2,
                  categories:
                    [
                    {
                      label: "test1",
                      value: "value1"
                    },
                    {
                      label: "test2",
                      value: "value2"
                    }
                    ]
                },
              description: "A trait",
              mainAbbreviation: "t1",
              abbreviations: ["t1","t2"],
              synonyms: ["stalk height"],
            }
            ]
          program:
            {
              id: "20e1a888-050a-4f89-8fec-ffaa55289c1b",
              name: "Nick's Microgreens",
              active: true,
              createdAt: "2020-06-24T11:36:01-04:00",
              updatedAt: "2020-06-24T11:36:01-04:00"
            }
          user:
            {
              id: "285ebacc-3940-416b-8ffe-5bb5d8a843f3",
              orcid: "0000-0000-0000-0000",
              name: "Nick Palladino",
              email: "n@n.com",
              active: true,
              systemRoles: []
            }
          createdByUser:
            {
              name: "Tester",
              active: true
            }
          updatedByUser:
            {
              name: "Tester",
              active: true
            }
...
  parameters:
    page:
      description: 'Used to request a specific page of data to be returned.
        The page indexing starts at 1 (the first page is ''page''= 1). Default is `1`.'
      example: 1
      in: query
      name: page
      required: false
      schema:
        type: integer
    pageSize:
      description: The size of the pages to be returned. Default is to show all results.
      example: 1000
      in: query
      name: pageSize
      required: false
      schema:
        type: integer
    sortOrder:
      description: Order in which to sort the results. Default is ascending order.
      example: ASC
      in: query
      name: sortOrder
      required: false
      schema:
        type: string